+++ packages/docs/pages/plugins/framekit.mdx
@@ -2,40 +2,12 @@
 
 MessageKit enables developers to create interactive elements (frames) inside messaging applications. These frames can be used for various purposes such as handling payments, displaying transaction receipts, managing conversations, and creating custom interactive interfaces.
 
-## Frame Types
-
-frames supports several types of frames:
-
-- `Framesv2`: Interact with agents via frames in Farcaster
 - `Payment Frames`: Request and handle cryptocurrency payments
 - `Receipt Frames`: Display transaction confirmations
 - `Conversation Frames`: Manage DMs and group messages
 - `Custom Frames`: Create custom interactive UI elements
 
-Below are practical examples of implementing each frame type.
-
-## Framesv2
-
-Framesv2 is a new way to interact with agents via frames. It allows you to have your own landing page for your agent and interact with it via frames in Farcaster.
-
-```bash
-Share in Farcaster (Framev2): https://frames.message-kit.org/dm/0x...
-```
-
-**Properties:**
-
-- `address`: The address of the agent. Could be converse profile name , ens name or address
-
-## Agent Wallet
-
-You can send agent wallet info using the `sendWallet` method:
-
-```typescript
-// Send agent wallet info
-await framekit.sendWallet("0x93E2fc3e99dFb1238eB9e0eF2580EFC5809C7204");
-```
-
-## Payment
+## Request payment
 
 You can request payments using the payment frame:
 
@@ -44,13 +16,22 @@ You can request payments using the payment frame:
 await framekit.requestPayment(recipientAddress, 1, "USDC");
 ```
 
+## Wallet details
+
+You can send agent wallet info using the `sendWallet` method:
+
+```typescript
+// Send agent wallet info
+await framekit.sendWallet("0x93E2fc3e99dFb1238eB9e0eF2580EFC5809C7204");
+```
+
 Properties:
 
 - `amount`: Number representing the payment amount
 - `token`: Supported tokens include "eth", "dai", "usdc", "degen"
 - `address`: Recipient's wallet address or ENS name
 
-## Transaction Receipts
+## Transfer receipt
 
 You can request receipts using the receipt frame:
 
@@ -109,6 +90,8 @@ Properties:
 - `action`: Type of button action ("link" or "post")
 - `target`: Destination URL for button clicks
 
+:::info
+
 ## Open Frames
 
 In compliance with [Open Frames](https://www.openframes.xyz/), use a meta tag in your frame's HTML to declare the client protocols your frame supports.
@@ -122,3 +105,5 @@ These are quickstarts for different Frameworks that support Open Frames
 - [OnchainKit Quickstart](https://github.com/ephemeraHQ/open-frames-starter-onchainkit/): Onchain Kit quickstart that integrates XMTP.
 - [Frog Quickstart](https://github.com/ephemeraHQ/open-frames-starter-frog): Frog XMTP quickstart
 - [Frames.js Quickstart](https://github.com/ephemeraHQ/open-frames-starter-framesjs/): Frames.js quickstart that integrates XMTP.
+
+:::
