+++ packages/docs/pages/skills/concierge.mdx
@@ -4,7 +4,7 @@ import Video from "../../components/Video";
 
 The Wallet Concierge acts as a personal financial assistant within your Agent. It handles the user financial needs without you needing to worry about the logic.
 
-<Video src="/concierge.mp4" />
+<Video src="/videos/concierge.mp4" />
 
 > As seen in this [tweet](https://x.com/xmtp_/status/1867254407819415653)
