+++ packages/docs/pages/ux/frames.mdx
@@ -1,0 +1,21 @@
+import Video from "../../components/Video";
+import { CustomHomePage } from "../../components/CustomHomePage";
+
+# Frames
+
+Frames are a way to interact with your agent via Farcaster. They are a way to interact with your agent via Farcaster.
+
+<Video src="/videos/frames.mp4" />
+
+## Framekit
+
+Framekit is a way to generate dynamic urls for your agent. They can work on Browsers or in a frame rendering environment.
+
+<CustomHomePage.TileGrid3>
+  <CustomHomePage.Tile
+    href="/plugins/framekit"
+    title="Framekit"
+    description="A kit for generating dynamic urls."
+    icon="🖼️"
+  />
+</CustomHomePage.TileGrid3>
