+++ packages/message-kit/tests/Frames.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect } from "vitest";
-import { extractFrameChain } from "../src/helpers/utils";
+import { extractFrameChain } from "../../framekit/src/app/utils/networks";
 
 describe("Frame Chain Tests", () => {
   test.each([
