+++ packages/message-kit/tests/Parsing.test.ts
@@ -60,4 +60,4 @@ describe("Parsing tests", () => {
       }
     },
   );
-});
+}, 15000);
